チーム管理もPJ情報に統合
