+++ src/app/project-info/page.tsx
@@ -11,8 +11,12 @@ import { useAuth } from '@/services/auth/hooks/useAuth'
 import Auth from '@/services/auth/components/Auth'
 import ReactMarkdown from 'react-markdown'
 import rehypeRaw from 'rehype-raw'
-import { IoPencil, IoEye, IoLogoGoogle, IoLogoDropbox, IoCloudOutline } from 'react-icons/io5'
+import { IoPencil, IoEye, IoLogoGoogle, IoLogoDropbox, IoCloudOutline, IoPeople } from 'react-icons/io5'
 import { SiNotion } from 'react-icons/si'
+import { FiPlus, FiX, FiUserCheck, FiEdit2, FiTrash2, FiMail, FiUsers, FiUpload, FiFile, FiCheckCircle, FiInfo, FiTag, FiChevronDown, FiChevronRight } from 'react-icons/fi'
+import { User } from '@/features/tasks/types/user'
+import { ProjectMember, ProjectMemberRole, Skill, SkillLevel } from '@/features/projects/types/projectMember'
+import { getAllUsers } from '@/utils/memberUtils'
 
 // 情報入力項目の定義
 interface CompletionItem {
@@ -20,11 +24,27 @@ interface CompletionItem {
   isComplete: boolean;
 }
 
+// メンバー表示用の拡張インターフェース
+interface MemberWithUser extends ProjectMember {
+  user: User
+  tags?: string[]
+  position?: { x: number, y: number } 
+}
+
+// 招待済みユーザーの型定義
+interface InvitedUser {
+  id: string;
+  email: string;
+  role: ProjectMemberRole;
+  status: 'pending' | 'accepted' | 'declined';
+  invitedAt: Date;
+}
+
 // タブの定義
-type TabType = 'project-info' | 'external-reference';
+type TabType = 'project-info' | 'external-reference' | 'team';
 
 export default function ProjectInfo() {
-  const { currentProject } = useProjectContext()
+  const { currentProject, getProjectMembers, assignUserToProject, removeUserFromProject } = useProjectContext()
   const { isAuthenticated, user, login, logout } = useAuth()
   const [activeTab, setActiveTab] = useState('project-info')
   const [activeContentTab, setActiveContentTab] = useState<TabType>('project-info')
@@ -39,6 +59,40 @@ export default function ProjectInfo() {
   const fileInputRef = useRef<HTMLInputElement>(null)
   const [completionRate, setCompletionRate] = useState(0)
   const [completionItems, setCompletionItems] = useState<CompletionItem[]>([])
+  
+  // チーム管理関連の状態変数
+  const [members, setMembers] = useState<MemberWithUser[]>([])
+  const [showAddModal, setShowAddModal] = useState(false)
+  const [showEditModal, setShowEditModal] = useState(false)
+  const [selectedMember, setSelectedMember] = useState<MemberWithUser | null>(null)
+  const [selectedRole, setSelectedRole] = useState<ProjectMemberRole>('member')
+  const [selectedUserId, setSelectedUserId] = useState<string>('')
+  const [allUsers, setAllUsers] = useState<User[]>([])
+  
+  // 招待関連の状態
+  const [addUserTab, setAddUserTab] = useState<'existing' | 'invite'>('existing')
+  const [inviteEmail, setInviteEmail] = useState('')
+  const [isInviting, setIsInviting] = useState(false)
+  const [inviteSuccess, setInviteSuccess] = useState(false)
+  const [invitedUsers, setInvitedUsers] = useState<InvitedUser[]>([])
+  
+  // スキル関連の状態
+  const [showSkillModal, setShowSkillModal] = useState(false)
+  const [memberSkills, setMemberSkills] = useState<Skill[]>([])
+  const [skillDescription, setSkillDescription] = useState('')
+  const [skillSheetFile, setSkillSheetFile] = useState<string | null>(null)
+  const [showSkillDetailModal, setShowSkillDetailModal] = useState(false)
+  const [newSkill, setNewSkill] = useState<Skill>({ name: '', level: 'intermediate' })
+  const [memberTags, setMemberTags] = useState<string[]>([])
+  const [newTag, setNewTag] = useState('')
+  const skillFileInputRef = useRef<HTMLInputElement>(null)
+  
+  // チーム表示モード
+  const [teamViewMode, setTeamViewMode] = useState<'list' | 'structure'>('list')
+  const [expandedMembers, setExpandedMembers] = useState<Set<string>>(new Set())
+  
+  // メール形式検証用の正規表現
+  const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/
 
   // プロジェクトが変更されたときに初期化
   useEffect(() => {
@@ -72,6 +126,258 @@ export default function ProjectInfo() {
       setCompletionItems(items)
     }
   }, [currentProject, isInitialized, originalText, summaryText])
+
+  // ユーザーリストを初期化
+  useEffect(() => {
+    const users = getAllUsers().map(u => ({
+      id: u.id,
+      name: u.name,
+      role: u.role
+    } as User))
+    setAllUsers(users)
+  }, [])
+
+  // プロジェクトが変更されたらメンバーリストを更新
+  useEffect(() => {
+    if (currentProject) {
+      updateMemberList()
+    }
+  }, [currentProject, allUsers])
+  
+  // メンバーリストを更新する関数
+  const updateMemberList = () => {
+    if (!currentProject) return
+    
+    const projectMembers = getProjectMembers(currentProject.id)
+    
+    // メンバー情報にユーザー情報を紐付け
+    const membersWithUsers = projectMembers
+      .map(member => {
+        const user = allUsers.find(u => u.id === member.userId)
+        if (!user) return null
+        const existingMember = members.find(m => m.id === member.id);
+        return {
+          ...member,
+          user,
+          // 位置情報を保持（既存のものがあれば）
+          position: existingMember?.position
+        } as MemberWithUser;
+      })
+      .filter((member): member is MemberWithUser => member !== null)
+    
+    setMembers(membersWithUsers)
+  }
+
+  // メンバー追加モーダルを開く
+  const openAddModal = () => {
+    setSelectedUserId('')
+    setSelectedRole('member')
+    setInviteEmail('')
+    setInviteSuccess(false)
+    setAddUserTab('existing')
+    setShowAddModal(true)
+  }
+  
+  // メンバー編集モーダルを開く
+  const openEditModal = (member: MemberWithUser) => {
+    setSelectedMember(member)
+    setSelectedRole(member.role)
+    setShowEditModal(true)
+  }
+  
+  // スキル詳細モーダルを開く
+  const openSkillDetailModal = (member: MemberWithUser) => {
+    setSelectedMember(member)
+    setMemberSkills(member.skills || [])
+    setSkillDescription(member.skillDescription || '')
+    setSkillSheetFile(member.skillSheetFile || null)
+    setMemberTags(member.tags || [])
+    setShowSkillDetailModal(true)
+  }
+  
+  // スキル編集モーダルを開く
+  const openSkillModal = (member: MemberWithUser) => {
+    setSelectedMember(member)
+    setMemberSkills(member.skills || [])
+    setSkillDescription(member.skillDescription || '')
+    setSkillSheetFile(member.skillSheetFile || null)
+    setMemberTags(member.tags || [])
+    setShowSkillModal(true)
+  }
+
+  // メンバーを追加
+  const handleAddMember = () => {
+    if (addUserTab === 'existing') {
+      if (!selectedUserId || !currentProject) return
+      
+      assignUserToProject(currentProject.id, selectedUserId, selectedRole)
+      setShowAddModal(false)
+      updateMemberList()
+    } else {
+      handleInviteUser()
+    }
+  }
+  
+  // 外部ユーザーを招待
+  const handleInviteUser = () => {
+    if (!inviteEmail || !emailRegex.test(inviteEmail) || !currentProject) return
+    
+    setIsInviting(true)
+    
+    // 招待メール送信のシミュレーション
+    setTimeout(() => {
+      // 招待レコードを作成
+      const newInvite: InvitedUser = {
+        id: `invite-${Date.now()}-${Math.random().toString(36).substr(2, 9)}`,
+        email: inviteEmail,
+        role: selectedRole,
+        status: 'pending',
+        invitedAt: new Date()
+      }
+      
+      // 招待リストに追加
+      setInvitedUsers(prev => [...prev, newInvite])
+      
+      // 状態をリセット
+      setInviteSuccess(true)
+      setIsInviting(false)
+      
+      // 3秒後にメッセージを非表示
+      setTimeout(() => {
+        setShowAddModal(false)
+        setInviteSuccess(false)
+        setInviteEmail('')
+      }, 3000)
+    }, 1500) // 1.5秒間の遅延でメール送信をシミュレート
+  }
+  
+  // メンバー情報を更新
+  const handleUpdateMember = () => {
+    if (!selectedMember || !currentProject) return
+    
+    assignUserToProject(currentProject.id, selectedMember.userId, selectedRole)
+    setShowEditModal(false)
+    updateMemberList()
+  }
+  
+  // メンバーを削除
+  const handleRemoveMember = (userId: string) => {
+    if (!currentProject) return
+    
+    if (confirm('このメンバーをプロジェクトから削除しますか？')) {
+      removeUserFromProject(currentProject.id, userId)
+      updateMemberList()
+    }
+  }
+  
+  // スキルを追加
+  const addSkill = () => {
+    if (!newSkill.name.trim()) return
+    
+    // 既存のスキルと重複チェック
+    if (memberSkills.some(s => s.name.toLowerCase() === newSkill.name.toLowerCase())) {
+      alert('既に同じ名前のスキルが登録されています')
+      return
+    }
+    
+    setMemberSkills([...memberSkills, { ...newSkill }])
+    setNewSkill({ name: '', level: 'intermediate' })
+  }
+  
+  // スキルを削除
+  const removeSkill = (index: number) => {
+    const updatedSkills = [...memberSkills]
+    updatedSkills.splice(index, 1)
+    setMemberSkills(updatedSkills)
+  }
+  
+  // タグを追加
+  const addTag = () => {
+    if (!newTag.trim()) return
+    
+    // 重複チェック
+    if (memberTags.includes(newTag.trim())) {
+      alert('既に同じタグが登録されています')
+      return
+    }
+    
+    setMemberTags([...memberTags, newTag.trim()])
+    setNewTag('')
+  }
+  
+  // タグを削除
+  const removeTag = (index: number) => {
+    const updatedTags = [...memberTags]
+    updatedTags.splice(index, 1)
+    setMemberTags(updatedTags)
+  }
+  
+  // スキルシートファイル選択ダイアログを開く
+  const handleFileSelect = () => {
+    if (skillFileInputRef.current) {
+      skillFileInputRef.current.click()
+    }
+  }
+  
+  // スキルシートファイル変更時の処理
+  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const file = e.target.files?.[0]
+    if (!file) return
+    
+    // ファイル名を保存（実際にはここでアップロードなどの処理を行う）
+    setSkillSheetFile(file.name)
+  }
+  
+  // スキル情報を保存
+  const saveSkillInfo = () => {
+    if (!selectedMember || !currentProject) return
+    
+    // メンバー情報を更新
+    // 実際には更新APIを呼び出すなどの処理が必要
+    const updatedMembers = members.map(member => {
+      if (member.id === selectedMember.id) {
+        return {
+          ...member,
+          skills: memberSkills,
+          skillDescription: skillDescription,
+          skillSheetFile: skillSheetFile,
+          tags: memberTags
+        } as MemberWithUser
+      }
+      return member
+    })
+    
+    setMembers(updatedMembers)
+    setShowSkillModal(false)
+  }
+  
+  // ロールの日本語表示
+  const getRoleLabel = (role: ProjectMemberRole) => {
+    switch (role) {
+      case 'manager':
+        return 'マネージャー'
+      case 'member':
+        return 'メンバー'
+      default:
+        return role
+    }
+  }
+  
+  // スキルレベルの日本語表示
+  const getSkillLevelLabel = (level: SkillLevel) => {
+    switch (level) {
+      case 'beginner':
+        return '初級'
+      case 'intermediate':
+        return '中級'
+      case 'advanced':
+        return '上級'
+      case 'expert':
+        return 'エキスパート'
+      default:
+        return level
+    }
+  }
 
   // プロジェクト情報の入力度を計算
   const calculateCompletionRate = (project: Project): { rate: number; items: CompletionItem[] } => {
@@ -334,6 +640,757 @@ ${project.risks || 'リスク・課題情報はまだ入力されていません
           </div>
         );
 
+      case 'team':
+        return (
+          <div className="p-6">
+            <div className="flex justify-between items-center mb-4">
+              <h3 className="text-lg font-medium text-gray-900">チーム管理</h3>
+              <div className="flex gap-2">
+                <button
+                  onClick={() => setTeamViewMode('list')}
+                  className={`px-3 py-1.5 text-xs font-medium rounded-md flex items-center gap-1 ${
+                    teamViewMode === 'list'
+                      ? 'bg-blue-600 text-white'
+                      : 'bg-gray-100 text-gray-700 hover:bg-gray-200'
+                  }`}
+                >
+                  <FiUsers className="w-3.5 h-3.5" />
+                  リスト表示
+                </button>
+                <button
+                  onClick={() => setTeamViewMode('structure')}
+                  className={`px-3 py-1.5 text-xs font-medium rounded-md flex items-center gap-1 ${
+                    teamViewMode === 'structure'
+                      ? 'bg-blue-600 text-white'
+                      : 'bg-gray-100 text-gray-700 hover:bg-gray-200'
+                  }`}
+                >
+                  <FiUsers className="w-3.5 h-3.5" />
+                  構成表示
+                </button>
+                <button
+                  onClick={openAddModal}
+                  className="px-3 py-1.5 text-xs font-medium rounded-md bg-green-600 text-white hover:bg-green-700 flex items-center gap-1"
+                >
+                  <FiPlus className="w-3.5 h-3.5" />
+                  メンバー追加
+                </button>
+              </div>
+            </div>
+
+            {teamViewMode === 'list' ? (
+              <div className="bg-white shadow overflow-hidden rounded-lg">
+                <table className="min-w-full divide-y divide-gray-200">
+                  <thead className="bg-gray-50">
+                    <tr>
+                      <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                        メンバー
+                      </th>
+                      <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                        役割
+                      </th>
+                      <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                        スキル情報
+                      </th>
+                      <th scope="col" className="px-6 py-3 text-right text-xs font-medium text-gray-500 uppercase tracking-wider">
+                        操作
+                      </th>
+                    </tr>
+                  </thead>
+                  <tbody className="bg-white divide-y divide-gray-200">
+                    {members.length === 0 ? (
+                      <tr>
+                        <td colSpan={4} className="px-6 py-4 text-center text-sm text-gray-500">
+                          メンバーがいません。「メンバー追加」ボタンからメンバーを追加してください。
+                        </td>
+                      </tr>
+                    ) : (
+                      members.map(member => (
+                        <tr key={member.id} className="hover:bg-gray-50">
+                          <td className="px-6 py-4 whitespace-nowrap">
+                            <div className="flex items-center">
+                              <div className="flex-shrink-0 h-10 w-10 bg-gray-200 rounded-full flex items-center justify-center">
+                                <span className="text-sm font-medium text-gray-600">
+                                  {member.user.name.substring(0, 2).toUpperCase()}
+                                </span>
+                              </div>
+                              <div className="ml-4">
+                                <div className="text-sm font-medium text-gray-900">{member.user.name}</div>
+                                <div className="text-sm text-gray-500">{member.user.role}</div>
+                              </div>
+                            </div>
+                          </td>
+                          <td className="px-6 py-4 whitespace-nowrap">
+                            <span className={`px-2 py-1 inline-flex text-xs leading-5 font-semibold rounded-full 
+                              ${member.role === 'manager' ? 'bg-green-100 text-green-800' : 'bg-blue-100 text-blue-800'}`}>
+                              {getRoleLabel(member.role)}
+                            </span>
+                          </td>
+                          <td className="px-6 py-4 whitespace-nowrap">
+                            {member.skills && member.skills.length > 0 ? (
+                              <div className="flex flex-wrap gap-1">
+                                {member.skills.slice(0, 2).map((skill, index) => (
+                                  <span key={index} className="inline-flex items-center px-2 py-0.5 rounded text-xs font-medium bg-gray-100 text-gray-800">
+                                    {skill.name}
+                                  </span>
+                                ))}
+                                {member.skills.length > 2 && (
+                                  <span className="inline-flex items-center px-2 py-0.5 rounded text-xs font-medium bg-gray-100 text-gray-800">
+                                    +{member.skills.length - 2}
+                                  </span>
+                                )}
+                              </div>
+                            ) : (
+                              <button
+                                onClick={() => openSkillModal(member)}
+                                className="inline-flex items-center px-2 py-0.5 rounded text-xs font-medium bg-gray-100 text-gray-800 hover:bg-gray-200"
+                              >
+                                スキル追加
+                              </button>
+                            )}
+                          </td>
+                          <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
+                            <div className="flex justify-end gap-2">
+                              <button
+                                onClick={() => openSkillDetailModal(member)}
+                                className="text-indigo-600 hover:text-indigo-900"
+                              >
+                                <FiInfo className="h-4 w-4" />
+                                <span className="sr-only">詳細</span>
+                              </button>
+                              <button
+                                onClick={() => openEditModal(member)}
+                                className="text-blue-600 hover:text-blue-900"
+                              >
+                                <FiEdit2 className="h-4 w-4" />
+                                <span className="sr-only">編集</span>
+                              </button>
+                              <button
+                                onClick={() => handleRemoveMember(member.userId)}
+                                className="text-red-600 hover:text-red-900"
+                              >
+                                <FiTrash2 className="h-4 w-4" />
+                                <span className="sr-only">削除</span>
+                              </button>
+                            </div>
+                          </td>
+                        </tr>
+                      ))
+                    )}
+                  </tbody>
+                </table>
+              </div>
+            ) : (
+              <div className="bg-white shadow rounded-lg p-6">
+                <div className="text-center text-gray-600 py-20">
+                  <FiUsers className="mx-auto h-12 w-12 text-gray-400" />
+                  <h3 className="mt-2 text-sm font-medium text-gray-900">チーム構成表示</h3>
+                  <p className="mt-1 text-sm text-gray-500">
+                    チームの構造を視覚的に確認できます。
+                  </p>
+                </div>
+              </div>
+            )}
+
+            {/* メンバー追加モーダル */}
+            {showAddModal && (
+              <div className="fixed inset-0 bg-gray-600 bg-opacity-50 flex items-center justify-center z-50">
+                <div className="bg-white rounded-lg shadow-xl max-w-md w-full p-6">
+                  <div className="flex justify-between items-center mb-4">
+                    <h3 className="text-lg font-medium text-gray-900">メンバー追加</h3>
+                    <button
+                      type="button"
+                      onClick={() => setShowAddModal(false)}
+                      className="text-gray-400 hover:text-gray-500"
+                    >
+                      <FiX className="h-5 w-5" />
+                    </button>
+                  </div>
+                  
+                  <div className="border-b border-gray-200 mb-4">
+                    <nav className="flex -mb-px space-x-8">
+                      <button
+                        onClick={() => setAddUserTab('existing')}
+                        className={`py-2 px-1 border-b-2 font-medium text-sm ${
+                          addUserTab === 'existing'
+                            ? 'border-blue-500 text-blue-600'
+                            : 'border-transparent text-gray-500 hover:text-gray-700 hover:border-gray-300'
+                        }`}
+                      >
+                        既存ユーザー
+                      </button>
+                      <button
+                        onClick={() => setAddUserTab('invite')}
+                        className={`py-2 px-1 border-b-2 font-medium text-sm ${
+                          addUserTab === 'invite'
+                            ? 'border-blue-500 text-blue-600'
+                            : 'border-transparent text-gray-500 hover:text-gray-700 hover:border-gray-300'
+                        }`}
+                      >
+                        招待
+                      </button>
+                    </nav>
+                  </div>
+                  
+                  {addUserTab === 'existing' ? (
+                    <div>
+                      <div className="mb-4">
+                        <label htmlFor="user" className="block text-sm font-medium text-gray-700">ユーザー</label>
+                        <select
+                          id="user"
+                          value={selectedUserId}
+                          onChange={(e) => setSelectedUserId(e.target.value)}
+                          className="mt-1 block w-full pl-3 pr-10 py-2 text-base border-gray-300 focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm rounded-md"
+                        >
+                          <option value="">ユーザーを選択</option>
+                          {allUsers
+                            .filter(u => !members.some(m => m.userId === u.id))
+                            .map(user => (
+                              <option key={user.id} value={user.id}>
+                                {user.name} ({user.role})
+                              </option>
+                            ))}
+                        </select>
+                      </div>
+                      
+                      <div className="mb-6">
+                        <label htmlFor="role" className="block text-sm font-medium text-gray-700">役割</label>
+                        <select
+                          id="role"
+                          value={selectedRole}
+                          onChange={(e) => setSelectedRole(e.target.value as ProjectMemberRole)}
+                          className="mt-1 block w-full pl-3 pr-10 py-2 text-base border-gray-300 focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm rounded-md"
+                        >
+                          <option value="member">メンバー</option>
+                          <option value="manager">マネージャー</option>
+                        </select>
+                      </div>
+                      
+                      <div className="flex justify-end">
+                        <button
+                          type="button"
+                          onClick={() => setShowAddModal(false)}
+                          className="bg-white py-2 px-4 border border-gray-300 rounded-md shadow-sm text-sm font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500"
+                        >
+                          キャンセル
+                        </button>
+                        <button
+                          type="button"
+                          onClick={handleAddMember}
+                          disabled={!selectedUserId}
+                          className="ml-3 inline-flex justify-center py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 disabled:bg-gray-300 disabled:cursor-not-allowed"
+                        >
+                          追加
+                        </button>
+                      </div>
+                    </div>
+                  ) : (
+                    <div>
+                      {inviteSuccess ? (
+                        <div className="rounded-md bg-green-50 p-4">
+                          <div className="flex">
+                            <div className="flex-shrink-0">
+                              <FiCheckCircle className="h-5 w-5 text-green-400" />
+                            </div>
+                            <div className="ml-3">
+                              <p className="text-sm font-medium text-green-800">
+                                招待メールを送信しました
+                              </p>
+                            </div>
+                          </div>
+                        </div>
+                      ) : (
+                        <>
+                          <div className="mb-4">
+                            <label htmlFor="email" className="block text-sm font-medium text-gray-700">メールアドレス</label>
+                            <input
+                              type="email"
+                              id="email"
+                              value={inviteEmail}
+                              onChange={(e) => setInviteEmail(e.target.value)}
+                              className="mt-1 block w-full border border-gray-300 rounded-md shadow-sm py-2 px-3 focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
+                              placeholder="招待するユーザーのメールアドレス"
+                            />
+                          </div>
+                          
+                          <div className="mb-6">
+                            <label htmlFor="invite-role" className="block text-sm font-medium text-gray-700">役割</label>
+                            <select
+                              id="invite-role"
+                              value={selectedRole}
+                              onChange={(e) => setSelectedRole(e.target.value as ProjectMemberRole)}
+                              className="mt-1 block w-full pl-3 pr-10 py-2 text-base border-gray-300 focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm rounded-md"
+                            >
+                              <option value="member">メンバー</option>
+                              <option value="manager">マネージャー</option>
+                            </select>
+                          </div>
+                          
+                          <div className="flex justify-end">
+                            <button
+                              type="button"
+                              onClick={() => setShowAddModal(false)}
+                              className="bg-white py-2 px-4 border border-gray-300 rounded-md shadow-sm text-sm font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500"
+                            >
+                              キャンセル
+                            </button>
+                            <button
+                              type="button"
+                              onClick={handleInviteUser}
+                              disabled={!inviteEmail || !emailRegex.test(inviteEmail) || isInviting}
+                              className="ml-3 inline-flex justify-center py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 disabled:bg-gray-300 disabled:cursor-not-allowed"
+                            >
+                              {isInviting ? (
+                                <>
+                                  <svg className="animate-spin -ml-1 mr-2 h-4 w-4 text-white" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
+                                    <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
+                                    <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
+                                  </svg>
+                                  送信中...
+                                </>
+                              ) : (
+                                '招待メール送信'
+                              )}
+                            </button>
+                          </div>
+                        </>
+                      )}
+                    </div>
+                  )}
+                </div>
+              </div>
+            )}
+
+            {/* メンバー編集モーダル */}
+            {showEditModal && selectedMember && (
+              <div className="fixed inset-0 bg-gray-600 bg-opacity-50 flex items-center justify-center z-50">
+                <div className="bg-white rounded-lg shadow-xl max-w-md w-full p-6">
+                  <div className="flex justify-between items-center mb-4">
+                    <h3 className="text-lg font-medium text-gray-900">メンバー編集</h3>
+                    <button
+                      type="button"
+                      onClick={() => setShowEditModal(false)}
+                      className="text-gray-400 hover:text-gray-500"
+                    >
+                      <FiX className="h-5 w-5" />
+                    </button>
+                  </div>
+                  
+                  <div className="mb-4">
+                    <div className="flex items-center mb-4">
+                      <div className="flex-shrink-0 h-12 w-12 bg-gray-200 rounded-full flex items-center justify-center">
+                        <span className="text-lg font-medium text-gray-600">
+                          {selectedMember.user.name.substring(0, 2).toUpperCase()}
+                        </span>
+                      </div>
+                      <div className="ml-4">
+                        <h4 className="text-lg font-medium text-gray-900">{selectedMember.user.name}</h4>
+                        <p className="text-sm text-gray-500">{selectedMember.user.role}</p>
+                      </div>
+                    </div>
+                    
+                    <div className="mb-6">
+                      <label htmlFor="edit-role" className="block text-sm font-medium text-gray-700">役割</label>
+                      <select
+                        id="edit-role"
+                        value={selectedRole}
+                        onChange={(e) => setSelectedRole(e.target.value as ProjectMemberRole)}
+                        className="mt-1 block w-full pl-3 pr-10 py-2 text-base border-gray-300 focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm rounded-md"
+                      >
+                        <option value="member">メンバー</option>
+                        <option value="manager">マネージャー</option>
+                      </select>
+                    </div>
+                  </div>
+                  
+                  <div className="flex justify-end">
+                    <button
+                      type="button"
+                      onClick={() => setShowEditModal(false)}
+                      className="bg-white py-2 px-4 border border-gray-300 rounded-md shadow-sm text-sm font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500"
+                    >
+                      キャンセル
+                    </button>
+                    <button
+                      type="button"
+                      onClick={handleUpdateMember}
+                      className="ml-3 inline-flex justify-center py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500"
+                    >
+                      更新
+                    </button>
+                  </div>
+                </div>
+              </div>
+            )}
+
+            {/* スキル詳細モーダル */}
+            {showSkillDetailModal && selectedMember && (
+              <div className="fixed inset-0 bg-gray-600 bg-opacity-50 flex items-center justify-center z-50">
+                <div className="bg-white rounded-lg shadow-xl max-w-md w-full p-6 max-h-[90vh] overflow-y-auto">
+                  <div className="flex justify-between items-center mb-4">
+                    <h3 className="text-lg font-medium text-gray-900">スキル詳細</h3>
+                    <button
+                      type="button"
+                      onClick={() => setShowSkillDetailModal(false)}
+                      className="text-gray-400 hover:text-gray-500"
+                    >
+                      <FiX className="h-5 w-5" />
+                    </button>
+                  </div>
+                  
+                  <div className="flex items-center mb-6">
+                    <div className="flex-shrink-0 h-12 w-12 bg-gray-200 rounded-full flex items-center justify-center">
+                      <span className="text-lg font-medium text-gray-600">
+                        {selectedMember.user.name.substring(0, 2).toUpperCase()}
+                      </span>
+                    </div>
+                    <div className="ml-4">
+                      <h4 className="text-lg font-medium text-gray-900">{selectedMember.user.name}</h4>
+                      <p className="text-sm text-gray-500">{getRoleLabel(selectedMember.role)}</p>
+                    </div>
+                  </div>
+                  
+                  <div className="mb-6">
+                    <h5 className="text-sm font-medium text-gray-700 mb-2">スキル一覧</h5>
+                    {memberSkills.length > 0 ? (
+                      <div className="space-y-2">
+                        {memberSkills.map((skill, index) => (
+                          <div key={index} className="flex items-center justify-between bg-gray-50 p-2 rounded">
+                            <div>
+                              <span className="text-sm font-medium text-gray-800">{skill.name}</span>
+                              <span className="ml-2 text-xs text-gray-500">{getSkillLevelLabel(skill.level)}</span>
+                            </div>
+                          </div>
+                        ))}
+                      </div>
+                    ) : (
+                      <p className="text-sm text-gray-500">スキル情報がありません</p>
+                    )}
+                  </div>
+                  
+                  {skillDescription && (
+                    <div className="mb-6">
+                      <h5 className="text-sm font-medium text-gray-700 mb-2">スキル詳細</h5>
+                      <p className="text-sm text-gray-600 whitespace-pre-line bg-gray-50 p-3 rounded">
+                        {skillDescription}
+                      </p>
+                    </div>
+                  )}
+                  
+                  {memberTags && memberTags.length > 0 && (
+                    <div className="mb-6">
+                      <h5 className="text-sm font-medium text-gray-700 mb-2">タグ</h5>
+                      <div className="flex flex-wrap gap-1">
+                        {memberTags.map((tag, index) => (
+                          <span key={index} className="inline-flex items-center px-2 py-0.5 rounded text-xs font-medium bg-gray-100 text-gray-800">
+                            {tag}
+                          </span>
+                        ))}
+                      </div>
+                    </div>
+                  )}
+                  
+                  {skillSheetFile && (
+                    <div className="mb-6">
+                      <h5 className="text-sm font-medium text-gray-700 mb-2">スキルシート</h5>
+                      <div className="flex items-center bg-gray-50 p-2 rounded">
+                        <FiFile className="h-5 w-5 text-gray-400 mr-2" />
+                        <span className="text-sm text-gray-600">{skillSheetFile}</span>
+                      </div>
+                    </div>
+                  )}
+                  
+                  <div className="flex justify-end">
+                    <button
+                      type="button"
+                      onClick={() => openSkillModal(selectedMember)}
+                      className="inline-flex justify-center py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500"
+                    >
+                      編集
+                    </button>
+                    <button
+                      type="button"
+                      onClick={() => setShowSkillDetailModal(false)}
+                      className="ml-3 inline-flex justify-center py-2 px-4 border border-gray-300 rounded-md shadow-sm text-sm font-medium text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500"
+                    >
+                      閉じる
+                    </button>
+                  </div>
+                </div>
+              </div>
+            )}
+
+            {/* スキル編集モーダル */}
+            {showSkillModal && selectedMember && (
+              <div className="fixed inset-0 bg-gray-600 bg-opacity-50 flex items-center justify-center z-50">
+                <div className="bg-white rounded-lg shadow-xl max-w-md w-full p-6 max-h-[90vh] overflow-y-auto">
+                  <div className="flex justify-between items-center mb-4">
+                    <h3 className="text-lg font-medium text-gray-900">スキル情報編集</h3>
+                    <button
+                      type="button"
+                      onClick={() => setShowSkillModal(false)}
+                      className="text-gray-400 hover:text-gray-500"
+                    >
+                      <FiX className="h-5 w-5" />
+                    </button>
+                  </div>
+                  
+                  <div className="mb-4">
+                    <div className="flex items-center mb-4">
+                      <div className="flex-shrink-0 h-12 w-12 bg-gray-200 rounded-full flex items-center justify-center">
+                        <span className="text-lg font-medium text-gray-600">
+                          {selectedMember.user.name.substring(0, 2).toUpperCase()}
+                        </span>
+                      </div>
+                      <div className="ml-4">
+                        <h4 className="text-lg font-medium text-gray-900">{selectedMember.user.name}</h4>
+                        <p className="text-sm text-gray-500">{getRoleLabel(selectedMember.role)}</p>
+                      </div>
+                    </div>
+                  </div>
+                  
+                  <div className="mb-6">
+                    <h5 className="text-sm font-medium text-gray-700 mb-2">スキル一覧</h5>
+                    
+                    <div className="space-y-2 mb-4">
+                      {memberSkills.map((skill, index) => (
+                        <div key={index} className="flex items-center justify-between bg-gray-50 p-2 rounded">
+                          <div>
+                            <span className="text-sm font-medium text-gray-800">{skill.name}</span>
+                            <span className="ml-2 text-xs text-gray-500">{getSkillLevelLabel(skill.level)}</span>
+                          </div>
+                          <button
+                            type="button"
+                            onClick={() => removeSkill(index)}
+                            className="text-red-500 hover:text-red-700"
+                          >
+                            <FiX className="h-4 w-4" />
+                          </button>
+                        </div>
+                      ))}
+                    </div>
+                    
+                    <div className="border-t border-gray-200 pt-4">
+                      <h5 className="text-sm font-medium text-gray-700 mb-2">スキル追加</h5>
+                      <div className="flex space-x-2 mb-2">
+                        <input
+                          type="text"
+                          value={newSkill.name}
+                          onChange={(e) => setNewSkill({ ...newSkill, name: e.target.value })}
+                          placeholder="スキル名"
+                          className="flex-1 block w-full border border-gray-300 rounded-md shadow-sm py-2 px-3 focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
+                        />
+                        <select
+                          value={newSkill.level}
+                          onChange={(e) => setNewSkill({ ...newSkill, level: e.target.value as SkillLevel })}
+                          className="block w-1/3 pl-3 pr-10 py-2 text-base border-gray-300 focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm rounded-md"
+                        >
+                          <option value="beginner">初級</option>
+                          <option value="intermediate">中級</option>
+                          <option value="advanced">上級</option>
+                          <option value="expert">エキスパート</option>
+                        </select>
+                      </div>
+                      <button
+                        type="button"
+                        onClick={addSkill}
+                        disabled={!newSkill.name}
+                        className="inline-flex justify-center py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 disabled:bg-gray-300 disabled:cursor-not-allowed"
+                      >
+                        スキル追加
+                      </button>
+                    </div>
+                  </div>
+                  
+                  <div className="mb-6">
+                    <label htmlFor="skill-description" className="block text-sm font-medium text-gray-700 mb-2">スキル詳細</label>
+                    <textarea
+                      id="skill-description"
+                      rows={4}
+                      value={skillDescription}
+                      onChange={(e) => setSkillDescription(e.target.value)}
+                      placeholder="スキルに関する詳細説明を記入してください。"
+                      className="block w-full border border-gray-300 rounded-md shadow-sm py-2 px-3 focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
+                    ></textarea>
+                  </div>
+                  
+                  <div className="mb-6">
+                    <h5 className="text-sm font-medium text-gray-700 mb-2">タグ</h5>
+                    
+                    <div className="flex flex-wrap gap-1 mb-2">
+                      {memberTags.map((tag, index) => (
+                        <div key={index} className="inline-flex items-center bg-gray-100 rounded px-2 py-1">
+                          <span className="text-xs text-gray-800 mr-1">{tag}</span>
+                          <button
+                            type="button"
+                            onClick={() => removeTag(index)}
+                            className="text-gray-500 hover:text-gray-700"
+                          >
+                            <FiX className="h-3 w-3" />
+                          </button>
+                        </div>
+                      ))}
+                    </div>
+                    
+                    <div className="flex space-x-2">
+                      <input
+                        type="text"
+                        value={newTag}
+                        onChange={(e) => setNewTag(e.target.value)}
+                        placeholder="新しいタグ"
+                        className="flex-1 block w-full border border-gray-300 rounded-md shadow-sm py-2 px-3 focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
+                        onKeyDown={(e) => e.key === 'Enter' && addTag()}
+                      />
+                      <button
+                        type="button"
+                        onClick={addTag}
+                        disabled={!newTag.trim()}
+                        className="inline-flex justify-center py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 disabled:bg-gray-300 disabled:cursor-not-allowed"
+                      >
+                        追加
+                      </button>
+                    </div>
+                  </div>
+                  
+                  <div className="mb-6">
+                    <label className="block text-sm font-medium text-gray-700 mb-2">スキルシート</label>
+                    <input
+                      type="file"
+                      ref={skillFileInputRef}
+                      onChange={handleFileChange}
+                      className="hidden"
+                      accept=".pdf,.doc,.docx,.xlsx"
+                    />
+                    
+                    {skillSheetFile ? (
+                      <div className="flex items-center justify-between bg-gray-50 p-2 rounded">
+                        <div className="flex items-center">
+                          <FiFile className="h-5 w-5 text-gray-400 mr-2" />
+                          <span className="text-sm text-gray-600">{skillSheetFile}</span>
+                        </div>
+                        <button
+                          type="button"
+                          onClick={() => setSkillSheetFile(null)}
+                          className="text-red-500 hover:text-red-700"
+                        >
+                          <FiX className="h-4 w-4" />
+                        </button>
+                      </div>
+                    ) : (
+                      <button
+                        type="button"
+                        onClick={handleFileSelect}
+                        className="inline-flex items-center px-4 py-2 border border-gray-300 shadow-sm text-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500"
+                      >
+                        <FiUpload className="mr-2 h-4 w-4" />
+                        ファイルをアップロード
+                      </button>
+                    )}
+                  </div>
+                  
+                  <div className="flex justify-end">
+                    <button
+                      type="button"
+                      onClick={() => setShowSkillModal(false)}
+                      className="bg-white py-2 px-4 border border-gray-300 rounded-md shadow-sm text-sm font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500"
+                    >
+                      キャンセル
+                    </button>
+                    <button
+                      type="button"
+                      onClick={saveSkillInfo}
+                      className="ml-3 inline-flex justify-center py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500"
+                    >
+                      保存
+                    </button>
+                  </div>
+                </div>
+              </div>
+            )}
+
+            {/* 招待済みユーザーリスト（オプション） */}
+            {invitedUsers.length > 0 && (
+              <div className="mt-8">
+                <h3 className="text-lg font-medium text-gray-900 mb-4">招待中のユーザー</h3>
+                <div className="bg-white shadow overflow-hidden rounded-lg">
+                  <table className="min-w-full divide-y divide-gray-200">
+                    <thead className="bg-gray-50">
+                      <tr>
+                        <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                          メールアドレス
+                        </th>
+                        <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                          役割
+                        </th>
+                        <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                          ステータス
+                        </th>
+                        <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                          招待日
+                        </th>
+                        <th scope="col" className="px-6 py-3 text-right text-xs font-medium text-gray-500 uppercase tracking-wider">
+                          操作
+                        </th>
+                      </tr>
+                    </thead>
+                    <tbody className="bg-white divide-y divide-gray-200">
+                      {invitedUsers.map(invite => (
+                        <tr key={invite.id} className="hover:bg-gray-50">
+                          <td className="px-6 py-4 whitespace-nowrap">
+                            <div className="flex items-center">
+                              <div className="flex-shrink-0 h-10 w-10 bg-gray-200 rounded-full flex items-center justify-center">
+                                <FiMail className="h-5 w-5 text-gray-500" />
+                              </div>
+                              <div className="ml-4">
+                                <div className="text-sm font-medium text-gray-900">{invite.email}</div>
+                              </div>
+                            </div>
+                          </td>
+                          <td className="px-6 py-4 whitespace-nowrap">
+                            <span className={`px-2 py-1 inline-flex text-xs leading-5 font-semibold rounded-full 
+                              ${invite.role === 'manager' ? 'bg-green-100 text-green-800' : 'bg-blue-100 text-blue-800'}`}>
+                              {getRoleLabel(invite.role)}
+                            </span>
+                          </td>
+                          <td className="px-6 py-4 whitespace-nowrap">
+                            <span className={`px-2 py-1 inline-flex text-xs leading-5 font-semibold rounded-full 
+                              ${
+                                invite.status === 'pending' 
+                                  ? 'bg-yellow-100 text-yellow-800' 
+                                  : invite.status === 'accepted' 
+                                    ? 'bg-green-100 text-green-800' 
+                                    : 'bg-red-100 text-red-800'
+                              }`}>
+                              {invite.status === 'pending' ? '招待中' : invite.status === 'accepted' ? '承諾済み' : '辞退'}
+                            </span>
+                          </td>
+                          <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
+                            {new Date(invite.invitedAt).toLocaleDateString('ja-JP')}
+                          </td>
+                          <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
+                            {invite.status === 'pending' && (
+                              <button
+                                onClick={() => {
+                                  if (confirm('この招待をキャンセルしますか？')) {
+                                    setInvitedUsers(prev => prev.filter(i => i.id !== invite.id))
+                                  }
+                                }}
+                                className="text-red-600 hover:text-red-900"
+                              >
+                                キャンセル
+                              </button>
+                            )}
+                          </td>
+                        </tr>
+                      ))}
+                    </tbody>
+                  </table>
+                </div>
+              </div>
+            )}
+          </div>
+        );
+
       default:
         return (
           <div className="flex flex-col md:flex-row gap-6 p-6">
@@ -478,6 +1535,17 @@ ${project.risks || 'リスク・課題情報はまだ入力されていません
               >
                 <IoCloudOutline className="mr-2" />
                 外部情報参照
+              </button>
+              <button
+                onClick={() => setActiveContentTab('team')}
+                className={`pb-4 px-1 border-b-2 font-medium text-sm inline-flex items-center ${
+                  activeContentTab === 'team'
+                    ? 'border-blue-500 text-blue-600'
+                    : 'border-transparent text-gray-500 hover:text-gray-700 hover:border-gray-300'
+                }`}
+              >
+                <IoPeople className="mr-2" />
+                チーム管理
               </button>
             </nav>
           </div>
